[WJ] Move tests to block.test.ts
